feat: move logic to common mixin to be used on Resources and UploadedFiles
