+++ solution/backend/resources/models.py
@@ -1,10 +1,7 @@
-from html import unescape
-
 from django.core.exceptions import NON_FIELD_ERRORS, ValidationError
 from django.db import models
 from django.db.models.signals import post_save
 from django.dispatch import receiver
-from django.utils.html import escape
 from django_jsonform.models.fields import ArrayField
 from model_utils.managers import InheritanceManager, InheritanceQuerySet
 from solo.models import SingletonModel
@@ -15,7 +12,7 @@ from common.fields import (
     UscRefField,
     VariableDateField,
 )
-from common.mixins import DisplayNameFieldMixin
+from common.mixins import DisplayNameFieldMixin, EscapeOnSaveMixin
 
 
 # Field mixins
@@ -26,24 +23,13 @@ class InternalNotesFieldMixin(models.Model):
         abstract = True
 
 
-class TypicalResourceFieldsMixin(InternalNotesFieldMixin):
+class TypicalResourceFieldsMixin(EscapeOnSaveMixin, InternalNotesFieldMixin):
     name = models.CharField(max_length=512, null=True, blank=True)
     description = models.TextField(null=True, blank=True)
     url = models.URLField(max_length=512, null=True, blank=True)
     name_sort = NaturalSortField('name', null=True)
     description_sort = NaturalSortField('description', null=True)
     date = VariableDateField(null=True, blank=True)
-
-    def save(self, *args, **kwargs):
-        # for loop through fields and escape fields that are CharFields or TextFields
-        for field in self._meta.fields:
-            if isinstance(field, (models.CharField, models.TextField)):
-                value = getattr(self, field.name)
-                if value:
-                    value = unescape(value)
-                    setattr(self, field.name, escape(value))
-
-        super(TypicalResourceFieldsMixin, self).save(*args, **kwargs)
 
     class Meta:
         abstract = True
